fix: check if user can unsubscribe

+++ juntagrico/views.py
@@ -91,7 +91,7 @@ def job(request, job_id):
                 # button does not trigger a resubmission of the form
                 return redirect('job', job_id=job_id)
 
-        elif request.POST.get('unsubscribe'):
+        elif request.POST.get('unsubscribe') and can_unsubscribe:
             assignment = Assignment.objects.filter(member=member, job=job)
             assignment.delete()
             # redirect to same page such that refresh in the browser or back
